requirements.txt
colorama==0.4.4
console_menu==0.6.0
consolemenu==1.0.1
PyAutoGUI==0.9.53
pyfiglet==0.8.post1
pypresence==4.2.0
python-dotenv==0.19.2
random_user_agent==1.0.1
requests==2.25.1
selenium==3.14.0
webdriver_auto_update==0.0.3.1
